apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
aiosmtplib==3.0.1
boto3==1.35.28
httpx==0.27.2
anyio==4.4.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
pillow==10.4.0
transformers==4.44.2
torch==2.3.1
ultralytics==8.3.24
numpy==1.26.4
python-docx==1.1.2